+++ src/optimizer/direct/DirectOptimizer.h
@@ -31,7 +31,7 @@ private:
 
     static functionValue estimatedValue(const shared_ptr<HyRect> &rect, double k);
 
-    list<shared_ptr<HyRect>> optimalRectangles(unsigned long m, functionValue phi);
+    list<shared_ptr<HyRect>> optimalRectangles(size_t m, functionValue phi);
 
     void addActiveRects(const list<shared_ptr<HyRect>> &rects);
