+++ src/optimizer/direct/StoppingCondition.h
@@ -11,8 +11,8 @@ using namespace std::chrono;
 
 class StoppingCondition {
 private:
-    const unsigned long NR_EVALUATIONS;
-    const unsigned long NR_HYRECTS;
+    const size_t NR_EVALUATIONS;
+    const size_t NR_HYRECTS;
 
     time_point<system_clock, seconds> END_TIME;
     const unsigned int mins;
@@ -26,12 +26,12 @@ private:
     bool updateAccuracy(functionValue newBestVal);
 
 public:
-    explicit StoppingCondition(unsigned long evaluations = 0, unsigned long hyrects = 0, unsigned int minutes = 0,
+    explicit StoppingCondition(size_t evaluations = 0, size_t hyrects = 0, unsigned int minutes = 0,
                                functionValue accuracy = 0, unsigned int accuracyIterations = 0);
 
     void setStartNow();
 
-    bool evaluate(unsigned long evaluations, unsigned long hyrects, functionValue newBestVal);
+    bool evaluate(size_t evaluations, size_t hyrects, functionValue newBestVal);
 };
